feat: #282 - 회원가입 방식 선택 및 공통 약관 흐름 개선
